Abstracted the ledger and ledger functions to their own requireable object. I made it so that the text on the end of series display can be changed to fit movies and shows that are currently airing.
Shows that are airing do not use the episode count provided my mal, it asks the php to return a count of the episodes. If the user gets to the end of a show that is not finished airing, it will show the end of series display with different text and not move the show to completed.